refactor(NotifyMessage): 단일 클래스로 패키지 삭제
